refactor: use redux environment_variables for EnvironmentVariables state
